// src/app/interfaces/rango-evaluativo.ts
export interface RangoEvaluativoInterface {
  idCampus: number;
  idCicloEscolar: number;
  idNivelAcademico: number;
  idRangoEvaluativo: number;
  nombreRangoEvaluativo: string;
  abreviatura: string;
  valorMinimo: string;
  valorMaximo: string;
  idColor: number;
  aprobado: boolean;
  activo: boolean;
}